Utility: generalize & test Windows error string helper file.
It will eventually contain also a stringifier for errno as well so it's
exposed for all platforms, but so far it's just that one Windows-only
function.

And the test fails, for some reason. Is it a stray \0 at the end coming
from the narrow() implementation? Let's see in the next commit.